core/requirements.txt
beautifulsoup4==4.11.1
bottle==0.12.23
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
clr-loader==0.2.5
cssselect2==0.6.0
cycler==0.11.0
fake-useragent==0.1.11
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
lxml==4.9.1
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
Pillow==9.1.0
proxy-tools==0.1.0
psutil==5.9.4
pycparser==2.21
pynput==1.7.6
pyparsing==3.0.9
python-dateutil==2.8.2
pythonnet==3.0.1
pywebview==4.0.1
QtPy==2.3.0
reportlab==3.6.10
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
svglib==1.3.0
svgutils==0.3.4
svgwrite==1.4.2
tinycss2==1.1.1
urllib3==1.26.8
webencodings==0.5.1
